Add files via upload
Java program to manage and control the seats that have been sold and the seats that are still available for one of their theatre sessions.
Asking user for option in the main menu and proceeds accordingly.
Main Menu consists of:
1. Buy ticket that asks the user to input a row number and seat number and check that the row and seat are correct and that the seat is available.
2. Print seating area that shows the seats that have been sold, and the seats that are still available.
3.Cancel ticket that makes a seat available again.
4.Show available that for each of the 3 rows displays the seats that are still available. 
5. Save that saves the 3 arrays with the row’s information in a file. 
6. Load that loads the file saved in previous task and restores the 3 arrays with the row’s information.

Additionally ,
Person (Person.java) with a constructor and the following attributes : name, surname, and email. Add a constructor that takes the 3 variables as input to create an object Person.
Ticket (Ticket.java) with a constructor and the following attributes: row, seat, price, and Person. Person will be is an object created using the class Person.
 method in class Ticket called print that prints all the information from a ticket: Person name, Person surname, Person email, row, seat, and price.

In the main program, an array list of tickets to save all the Tickets. Extend the buy_ticket method such that when buying a ticket, it asks for the information of a Person, creates a new ticket and adds the ticket in the new array list of tickets. Extend the cancel_ticket method such that when cancelling a ticket, it removes the ticket from the array list of tickets.
method called show_tickets_info that prints all the information for all tickets and calculates and shows the total price of all tickets after the ticket’s info. 
method called sort_tickets that returns a new list of Tickets sorted ticket prices in ascending order (cheapest first)- Used Bubble sort.

This Program asks the staff for each price and then proceeds further to ask from user multiple inputs as possible!
and finally displays all the tickets.